object-detector-v1/pyproject.toml
[tool.poetry]
name = "object-detector-v1"
version = "0.1.0"
description = ""
authors = ["Raphael Tang <raphpb1912@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
tensorflow = "^2.5.0"
Pillow = "^8.2.0"
numpy = "1.19.3"
matplotlib = "^3.4.2"
tensorflow-hub = "^0.12.0"

[tool.poetry.dev-dependencies]
black = {version = "^21.6b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
